+++ ZeShmouttsAssets/Blizzard API for Unity/Scripts/WoW Game Data API/Mythic Keystone Leaderboard API/MythicKeystoneLeaderboard.cs
@@ -18,15 +18,15 @@ namespace ZeShmouttsAssets.BlizzardAPI
 			/// <summary>
 			/// Coroutine that retrieves a WoW weekly Mythic Keystone leaderboard by period.
 			/// </summary>
+			/// <param name="region">The region of the data to retrieve.</param>
 			/// <param name="connectedRealmId">The ID of the connected realm.</param>
 			/// <param name="dungeonId">The ID of the dungeon.</param>
 			/// <param name="period">The unique identifier for the leaderboard period.</param>
 			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
 			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
 			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
-			/// <param name="region">The region of the data to retrieve.</param>
 			/// <returns></returns>
-			public static IEnumerator GetMythicKeystoneLeaderboard(int connectedRealmId, int dungeonId, int period, Action<WowMythicKeystoneLeaderboard_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = BattleNetRegion.UnitedStates)
+			public static IEnumerator GetMythicKeystoneLeaderboard(BattleNetRegion region, int connectedRealmId, int dungeonId, int period, Action<WowMythicKeystoneLeaderboard_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null)
 			{
 				string path = string.Format("/data/wow/connected-realm/{0}/mythic-leaderboard/{1}/period/{2}", connectedRealmId, dungeonId, period);
 				yield return SendRequest(region, namespaceDynamic, path, action_Result, ifModifiedSince, action_LastModified);