+++ .github/CHANGELOG.md
@@ -8,9 +8,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
  - Added an editor-only window to quickly retrieve an OAuth access token (menu : `Blizzard API/Get Access Token`)
    - This is mostly useful in Blizzard's documentation, by providing an access token to test endpoints
   
+## [0.27.0] - 2020-04-26
+### Added
+ - Added WoW Game Data endpoint support :
+   - Mythic Raid Leaderboard
+
+### Changed
+ - Changed the region parameter to mandatory for the Mythic Keystone Leaderboard endpoint
+  
 ## [0.26.0] - 2020-04-26
 ### Added
- - Added WoW Game Data endpoints support :
+ - Added WoW Game Data endpoint support :
    - Mythic Keystone Leaderboard
 
 ### Changed